requirements.txt
pyqt6
binance-connector
matplotlib
requests
websockets
